Refactor utilities and fix adapter functions
