+++ zw_mcp/blender_adapter.py
@@ -5,7 +5,8 @@ from pathlib import Path
 import argparse
 import math  # Added for math.radians
 from mathutils import Vector, Euler  # For ZW-COMPOSE transforms
-import ast
+
+from zw_mcp.utils import safe_eval, parse_color
 
 # Standardized Prefixes
 P_INFO = "[ZW->Blender][INFO]"
@@ -124,44 +125,6 @@ except ImportError:
 ZW_INPUT_FILE_PATH = Path("zw_mcp/prompts/blender_scene.zw")  # Default, can be overridden by args
 
 # --- Utility Functions ---
-def safe_eval(str_val, default_val):
-    """Safely evaluate a string to a Python literal."""
-    if not isinstance(str_val, str):
-        return default_val
-    try:
-        return ast.literal_eval(str_val)
-    except (ValueError, SyntaxError):
-        print(f"{P_WARN} safe_eval could not parse '{str_val}'. Using default {default_val}.")
-        return default_val
-
-
-def parse_color(color_val, default=(1.0, 1.0, 1.0, 1.0)):
-    """Parse a color definition from hex or tuple string formats."""
-    if not isinstance(color_val, str):
-        return default
-    val = color_val.strip()
-    if val.startswith("#"):
-        try:
-            r = int(val[1:3], 16) / 255.0
-            g = int(val[3:5], 16) / 255.0
-            b = int(val[5:7], 16) / 255.0
-            a = int(val[7:9], 16) / 255.0 if len(val) == 9 else 1.0
-            return (r, g, b, a)
-        except Exception:
-            return default
-    if val.startswith("(") and val.endswith(")"):
-        try:
-            tup = ast.literal_eval(val)
-            if isinstance(tup, (list, tuple)):
-                if len(tup) == 3:
-                    return (float(tup[0]), float(tup[1]), float(tup[2]), 1.0)
-                if len(tup) == 4:
-                    return (float(tup[0]), float(tup[1]), float(tup[2]), float(tup[3]))
-        except Exception:
-            return default
-    return default
-
-
 def get_or_create_collection(name: str, parent_collection=None):
     """Retrieve or create a Blender collection by name."""
     if not bpy:
